refactor(tests): split BYOK security tests under the 800-line guardrail
Move the 1749-line module into concern-grouped files sharing isolation
fixtures, without changing assertions. Collect count stays at 105.
