+++ backend/tests/fast_unit_duration_allowlist.txt
@@ -116,12 +116,12 @@ tests/unit/test_backend_runtime_env_validator.py::test_repo_prod_cloud_run_workf
 tests/unit/test_backend_runtime_env_validator.py::test_repo_prod_gke_values_match_manifest
 tests/unit/test_backend_runtime_env_validator.py::test_repo_rendered_cloud_run_matches_manifest
 tests/unit/test_verify_llm_gateway_serving.py::test_gateway_promotion_intent_tracks_runtime_and_helm_listener_surfaces
-tests/unit/test_byok_security.py::TestBYOKActivationValidation::test_valid_activation_persists_fingerprints
+tests/unit/test_byok_security_enrollment.py::TestBYOKActivationValidation::test_valid_activation_persists_fingerprints
 tests/unit/test_verify_pusher_config_references.py::test_dev_pusher_contract_requires_typesense_host_secret_clear
 tests/unit/test_verify_pusher_config_references.py::test_rendered_dev_pusher_direct_bindings_match_source_contract
-tests/unit/test_byok_security.py::TestCacheKeyHashing::test_hash_key_returns_sha256
-tests/unit/test_byok_security.py::TestChatQuotaBYOKBypass::test_enforce_chat_quota_bypasses_for_byok_with_openai_key
-tests/unit/test_byok_security.py::TestContextVarSafety::test_default_is_none
+tests/unit/test_byok_security_credentials.py::TestCacheKeyHashing::test_hash_key_returns_sha256
+tests/unit/test_byok_security_quota.py::TestChatQuotaBYOKBypass::test_enforce_chat_quota_bypasses_for_byok_with_openai_key
+tests/unit/test_byok_security_context.py::TestContextVarSafety::test_default_is_none
 tests/unit/test_byok_llm_logging.py::test_classify_byok_llm_error_authentication
 tests/unit/test_canonical_short_term_maintenance_cron.py::test_cohort_runner_noops_when_cron_disabled
 tests/unit/test_canonical_short_term_maintenance_cron.py::test_cohort_runner_uses_real_maintenance_with_fake_firestore
@@ -298,7 +298,7 @@ tests/unit/test_inv_mem_1_guard.py::TestInvMemSourceRatchet::test_no_non_canonic
 tests/unit/test_inv_mem_1_guard.py::TestInvMemSourceRatchet::test_default_read_paths_do_not_include_archive_without_explicit_mode
 tests/unit/test_task_intelligence_contract_freeze.py
 tests/unit/test_lock_bypass_fixes.py
-tests/unit/test_byok_security.py::TestCacheRouting::test_cached_openai_chat_no_raw_key_in_cache
+tests/unit/test_byok_security_enrollment.py::TestCacheRouting::test_cached_openai_chat_no_raw_key_in_cache
 # Behavior test drives the real process_segment target-attach fallback (#10033/#10119 review ask);
 # it is the sole heavy test in its file, so it amortizes the sync pipeline import into its call phase.
 tests/unit/test_sync_merge_target_selection.py::test_sync_target_attach_fallback_to_closest
